create mockbridge for invariant tests
this is a stripped down version of the bridge, with only the 4 functions
that we care about plus required members

we use this bridge to avoid vm.etch() which seems to cause forge to
start calling the functions of the bridge, even with excludeContract set

NOTE: had to hardcode addresses in the mockbridge - running forge test fails because the bridge's state is "empty"! when it shouldn't be, even tried passing the data throught the ctor instead of init. ultimately hardcoding the address in the bridge makes this works. clearly there's something wrong with forge

output from console.logs
  ....... deploying mock bridge
  hey! 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48 --> this confirms
  that the bridge has the data at post-deployment
  ....... deployed mock bridge
  Bound Result 4
  -------- convert
  Bound Result 6
  -------- deposit
  Bound Result 2
  -------- migrate -> this is the first call to migrate by forge, and
  the data is in the bridge!
  yyyyyyy fork 1
  xxxxxxx token 0xA8CE8aee21bC2A48a5EF670afCc9274C7bbbC035
  xxxxxxx tokenInfo origin token addr 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48
  xxxxxxx tokenInfo origin network 0
  Bound Result 2
  -------- migrate -> this is the second call, a repeat call by forge
  (why!?), and it says the data is not in the bridge???
  yyyyyyy fork 1
  xxxxxxx token 0xA8CE8aee21bC2A48a5EF670afCc9274C7bbbC035
  xxxxxxx tokenInfo origin token addr 0x0000000000000000000000000000000000000000
  xxxxxxx tokenInfo origin network 0

NOTE2: from the prints, I can see some asserts failing in the TEST!, but they do not cause the tests to revert (even with fail_on_revert = true); since this is happening after the migrate handler, it's most likely from the state problem described above