add separate env vars for the l1 owner + admin and the l2 owner + admin.
Prior to this commit, the deploy script assumed the owner + admin of the L1 and L2 contracts was the same address. This is not always the case,
so we add 2 more env vars here and update the deploy script to allow for the L2 and L1 contracts to be owned by separate addresses.